requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.2
protobuf==4.21.12
Werkzeug==2.3.8
WTForms==3.0.1
